Rename files back for better diffs
