+++ sharktank/sharktank/layers/__init__.py
@@ -10,7 +10,7 @@ from .paged_attention import PagedAttention, attn_type_map
 from .causal_llm import BaseCausalLMModel
 from .linear import LinearLayer
 from .norm import RMSNormLayer, LayerNorm
-from .rotary_embedding_sharded import build_rotary_layer
+from .rotary_embedding import build_rotary_layer
 from .token_embedding import TokenEmbeddingLayer
 from .paged_llama_attention_block import PagedLlamaAttentionBlock
 from .ffn_block import FFN
