+++ sharktank/sharktank/layers/latent_attention_block.py
@@ -10,7 +10,7 @@ from sharktank.types import *
 from .base import Theta, ThetaLayer
 from .linear import LinearLayer
 from .norm import RMSNormLayer
-from .rotary_embedding_sharded import ShardedRotaryLayer
+from .rotary_embedding import ShardedRotaryLayer
 from sharktank import ops
 
 __all__ = [
